requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
opencv-python
opencv-python-headless
torch
pillow
numpy